Max power is given as a float, not an int.

+++ lasers/cobalt.py
@@ -86,7 +86,7 @@ class CobaltLaser:
         # 'gmlp?' gets the maximum laser power in mW.
         self.write('gmlp?')
         response = self.readline()
-        return int(response)
+        return float(response)
 
 
     def getPower_mW(self):
